LPS-135843 Extract functions to its own file
